Trim the characters
